Extract shared chat mention chip
